requirements.txt
appdirs
astroid
black
certifi
click
colorama
cycler
distlib
filelock
kiwisolver
lazy-object-proxy
matplotlib
mccabe
mypy-extensions
numpy
opencv-python
pathspec
Pillow
pycparser
pylint
pyparsing
python-dateutil
pythonnet
regex
setuptools
six
toml
ast
typing-extensions
virtualenv
wrapt
